requirements.txt
dotenv
Flask
google-api-python-client
google-cloud-firestore
plotly
pytz